2008-12-16  Ole Reinhardt <ole.reinhardt@thermotemp.de> 	* gpioa_at91.c, gpiob_at91.c, gpioc_at91.c: 	  Fixed irq enable / disable bug
git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@2359 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ arch/arm/dev/gpioa_at91.c
@@ -75,7 +75,7 @@ static int PioCtlA(int cmd, void *param, int bit)
 {
     int rc = 0;
     u_int *ival = (u_int *) param;
-    uint16_t enabled = inr(PIOA_IMR) & _BV(bit);
+    uint32_t enabled = inr(PIOA_IMR) & _BV(bit);
 
     /* Disable interrupt. */
     if (enabled) {
