+++ arch/arm/dev/gpiob_at91.c
@@ -65,7 +65,7 @@ static int PioCtlB(int cmd, void *param, int bit)
 {
     int rc = 0;
     u_int *ival = (u_int *) param;
-    uint16_t enabled = inr(PIOB_IMR) & _BV(bit);
+    uint32_t enabled = inr(PIOB_IMR) & _BV(bit);
 
     /* Disable interrupt. */
     if (enabled) {
